1872 - Consolidate scattered file/S3 helpers into utils/file_utils.py
Gary's review on PR #1462 (ticket 1868) flagged that split_s3_uri had
ended up duplicated three times because neither existing utils.py is
safe to import from for the other framework's consumers (utils/utils.py
pulls in pyspark/pydeequ, polars_utils/utils.py pulls in polars).
Extracted every duplicated file/S3/CSV helper into one dependency-light
utils/file_utils.py that both sides can import safely, repointed every
consumer (the 4 Glue ingest jobs, cqc_api, postcode_matcher), deleted the
now-empty projects/_01_ingest/utils/utils.py, and added the missing
Dockerfile COPY for utils/file_utils.py to every Fargate image that
depends on it via polars_utils.
