+++ projects/_01_ingest/ascwds/jobs/ingest_ascwds_dataset.py
@@ -7,8 +7,7 @@ os.environ["SPARK_VERSION"] = "3.5"
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 
-from projects._01_ingest.utils import utils as iUtils
-from utils import utils
+from utils import file_utils, utils
 from utils.column_names.raw_data_files.ascwds_worker_columns import (
     AscwdsWorkerColumns as AWK,
 )
@@ -26,9 +25,9 @@ def main(source: str, destination: str, dataset: str = "ascwds"):
         destination (str): The destination directory for outputting parquet files.
         dataset (str): The dataset type, either 'ascwds' or 'nmdssc'. Defaults to 'ascwds'.
     """
-    bucket, prefix = utils.split_s3_uri(source)
+    bucket, prefix = file_utils.split_s3_uri(source)
 
-    if iUtils.is_csv(source):
+    if file_utils.is_csv(source):
         ingest_single_file(source, bucket, prefix, destination, dataset)
     else:
         ingest_multiple_files(bucket, prefix, destination, dataset)
@@ -48,7 +47,7 @@ def ingest_single_file(
         dataset (str): The dataset type, either 'ascwds' or 'nmdssc'.
     """
     print("Single file provided to job. Handling single file.")
-    new_destination = iUtils.construct_destination_path(destination, prefix)
+    new_destination = file_utils.construct_destination_path(destination, prefix)
     handle_job(source, bucket, prefix, new_destination, dataset)
 
 
@@ -63,14 +62,14 @@ def ingest_multiple_files(bucket: str, prefix: str, destination: str, dataset: s
         dataset (str): The dataset type, either 'ascwds' or 'nmdssc'.
     """
     print("Multiple files provided to job. Handling each file...")
-    objects_list = iUtils.get_s3_objects_list(bucket, prefix)
+    objects_list = file_utils.get_s3_objects_list(bucket, prefix)
 
     print("Objects list:")
     print(objects_list)
 
     for key in objects_list:
-        new_source = iUtils.construct_s3_uri(bucket, key)
-        new_destination = iUtils.construct_destination_path(destination, key)
+        new_source = file_utils.construct_s3_uri(bucket, key)
+        new_destination = file_utils.construct_destination_path(destination, key)
         handle_job(new_source, bucket, key, new_destination, dataset)
 
 
@@ -94,8 +93,8 @@ def handle_job(
     Raises:
         ValueError: If chosen method does not match 'ascwds' or 'nmdssc'.
     """
-    file_sample = iUtils.read_partial_csv_content(source_bucket, source_key)
-    delimiter = iUtils.identify_csv_delimiter(file_sample)
+    file_sample = file_utils.read_partial_csv_content(source_bucket, source_key)
+    delimiter = file_utils.identify_csv_delimiter(file_sample)
 
     df = utils.read_csv(source, delimiter)
     if dataset == "ascwds":
