+++ projects/_01_ingest/cqc_api/fargate/Dockerfile
@@ -8,6 +8,7 @@ ENV PYTHONPATH=/app
 
 # Dependencies
 COPY polars_utils polars_utils
+COPY utils/file_utils.py utils/file_utils.py
 COPY projects/_01_ingest/cqc_api/fargate/utils projects/_01_ingest/cqc_api/fargate/utils
 COPY projects/_01_ingest/cqc_api/utils/cqc_api.py projects/_01_ingest/cqc_api/utils/cqc_api.py
 COPY projects/_01_ingest/cqc_api/utils/validate_cqc_locations.py projects/_01_ingest/cqc_api/utils/validate_cqc_locations.py
