+++ projects/_01_ingest/ons_pd/jobs/ingest_ons_data.py
@@ -3,36 +3,35 @@ import sys
 
 os.environ["SPARK_VERSION"] = "3.5"
 
-from projects._01_ingest.utils import utils as iUtils
-from utils import utils
+from utils import file_utils, utils
 
 
 def main(source, destination):
-    if iUtils.is_csv(source):
+    if file_utils.is_csv(source):
         print("Single file provided to job. Handling single file.")
-        bucket, key = utils.split_s3_uri(source)
+        bucket, key = file_utils.split_s3_uri(source)
         print(destination)
-        new_destination = iUtils.construct_destination_path(destination, key)
+        new_destination = file_utils.construct_destination_path(destination, key)
         print(new_destination)
         handle_job(source, bucket, key, new_destination)
         return
 
     print("Multiple files provided to job. Handling each file...")
-    bucket, prefix = utils.split_s3_uri(source)
-    objects_list = iUtils.get_s3_objects_list(bucket, prefix)
+    bucket, prefix = file_utils.split_s3_uri(source)
+    objects_list = file_utils.get_s3_objects_list(bucket, prefix)
 
     print("Objects list:")
     print(objects_list)
 
     for key in objects_list:
-        new_source = iUtils.construct_s3_uri(bucket, key)
-        new_destination = iUtils.construct_destination_path(destination, key)
+        new_source = file_utils.construct_s3_uri(bucket, key)
+        new_destination = file_utils.construct_destination_path(destination, key)
         handle_job(new_source, bucket, key, new_destination)
 
 
 def handle_job(source: str, source_bucket: str, source_key: str, destination: str):
-    file_sample = iUtils.read_partial_csv_content(source_bucket, source_key)
-    delimiter = iUtils.identify_csv_delimiter(file_sample)
+    file_sample = file_utils.read_partial_csv_content(source_bucket, source_key)
+    delimiter = file_utils.identify_csv_delimiter(file_sample)
     ingest_dataset(source, destination, delimiter)
