+++ projects/_03_independent_cqc/_04_model/fargate/Dockerfile
@@ -8,6 +8,7 @@ ENV PYTHONPATH=/app
 
 # Dependencies
 COPY polars_utils /app/polars_utils
+COPY utils/file_utils.py /app/utils/file_utils.py
 COPY projects/_03_independent_cqc/_04_model/utils /app/projects/_03_independent_cqc/_04_model/utils
 COPY projects/_03_independent_cqc/_04_model/registry/model_registry.py /app/projects/_03_independent_cqc/_04_model/registry/model_registry.py
 COPY utils/column_names /app/utils/column_names/
