+++ projects/_04_direct_payment_recipients/Dockerfile_and_requirements/Dockerfile
@@ -10,6 +10,7 @@ ENV PYTHONPATH=/app
 COPY projects/_04_direct_payment_recipients/fargate/utils /app/projects/_04_direct_payment_recipients/fargate/utils
 COPY projects/_04_direct_payment_recipients/direct_payments_config_polars.py /app/projects/_04_direct_payment_recipients/direct_payments_config_polars.py
 COPY polars_utils /app/polars_utils
+COPY utils/file_utils.py /app/utils/file_utils.py
 COPY utils/column_names /app/utils/column_names/
 COPY utils/column_values /app/utils/column_values/
 COPY utils/value_labels /app/utils/value_labels
